Fixed SonarCloud issue

+++ src/raspberrypi/devices/primary_device.cpp
@@ -65,7 +65,7 @@ void PrimaryDevice::Inquiry()
 		LOGTRACE("Reporting LUN %d for device ID %d as not supported", lun, GetId())
 
 		// Signal that the requested LUN does not exist
-		controller->GetBuffer()[0] |= 0x7f;
+		controller->GetBuffer()[0] |= (int)0x7f;
 	}
 
 	EnterDataInPhase();
